code/requirements.txt
scikit_learn==0.24.2
python-docx
bs4
